Updated anomaly package from 4.0.2 to 4.3.0 and fixed plot_artifacts bug
Module `anomaly` was upgraded; `capa.mv` and `capa.uv` methods are now both simply capa.
The `transform` argument of `capa` was removed; sometimes, artifact detection
gave an execution error with the "check the `transform` function" flag.
I expect this bug to be resolved now?

There was a minor bug in plot_artifacts and it was fixed.